Accidently started a new branch from Animation start, so deleted those changes.
Created a QR-scanner and reader in LoginActivity.java. The security check to bypass the activity has not been implemented.

Created permissions in Manifest to use camera.

Created a new button in activity_login.xml to go to next Activity, since we dont want to scan a QRkode each time.

Inserted maven() googleurl in build.gradle.
Changed implementation to compile in build.gradle.